Modified folder structure and added CSV data files
