+++ solution/health_and_education/data/ref1_low_edu_gender_parity.csv
@@ -1,0 +1,48 @@
+Year,World,OECD90,Eastern Europe,Asia (Sans Japan),Middle East and Africa,Latin America,China,India,EU,USA
+2014,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0
+2015,3104.47059,0.0,8.48186,2065.8178,852.00259,178.16835,0.0,0.0,0.0,0.0
+2016,3161.44344,0.0,8.6802,2095.03147,876.99695,180.73482,0.0,0.0,0.0,0.0
+2017,3220.03211,0.0,8.88431,2124.97121,902.809,183.3676,0.0,0.0,0.0,0.0
+2018,3280.01976,0.0,9.09242,2155.4722,929.40282,186.05232,0.0,0.0,0.0,0.0
+2019,3341.07264,0.0,9.30224,2186.2827,956.72112,188.76657,0.0,0.0,0.0,0.0
+2020,3402.9354,0.0,9.51196,2217.20758,984.72307,191.4928,0.0,0.0,0.0,0.0
+2021,3465.51278,0.0,9.72091,2248.17352,1013.39323,194.22513,0.0,0.0,0.0,0.0
+2022,3528.83717,0.0,9.92908,2279.19598,1042.74746,196.96466,0.0,0.0,0.0,0.0
+2023,3592.92607,0.0,10.13609,2310.26903,1072.81245,199.70851,0.0,0.0,0.0,0.0
+2024,3657.83624,0.0,10.34174,2341.41128,1103.62806,202.45516,0.0,0.0,0.0,0.0
+2025,3723.5835,0.0,10.5459,2372.61663,1135.21914,205.20184,0.0,0.0,0.0,0.0
+2026,3790.16283,0.0,10.74844,2403.87092,1167.59705,207.94643,0.0,0.0,0.0,0.0
+2027,3857.46641,0.0,10.9493,2435.09384,1200.74062,210.68266,0.0,0.0,0.0,0.0
+2028,3925.2866,0.0,11.14859,2466.1362,1234.60293,213.39888,0.0,0.0,0.0,0.0
+2029,3993.35721,0.0,11.34656,2496.81047,1269.11927,216.08091,0.0,0.0,0.0,0.0
+2030,4061.48558,0.0,11.54352,2526.98012,1304.24345,218.71849,0.0,0.0,0.0,0.0
+2031,4129.59718,0.0,11.73953,2556.5916,1339.95944,221.30662,0.0,0.0,0.0,0.0
+2032,4197.73264,0.0,11.935,2585.66899,1376.28164,223.84701,0.0,0.0,0.0,0.0
+2033,4265.96225,0.0,12.13112,2614.25436,1413.23206,226.34472,0.0,0.0,0.0,0.0
+2034,4334.41009,0.0,12.32938,2642.42615,1450.84617,228.80839,0.0,0.0,0.0,0.0
+2035,4403.18004,0.0,12.53094,2670.25277,1489.15146,231.24487,0.0,0.0,0.0,0.0
+2036,4472.28136,0.0,12.73618,2697.73783,1528.15208,233.65528,0.0,0.0,0.0,0.0
+2037,4541.71837,0.0,12.9451,2724.8905,1567.84374,236.03903,0.0,0.0,0.0,0.0
+2038,4611.58966,0.0,13.15806,2751.79605,1608.23606,238.39949,0.0,0.0,0.0,0.0
+2039,4682.0159,0.0,13.37532,2778.56113,1649.33899,240.74046,0.0,0.0,0.0,0.0
+2040,4753.08948,0.0,13.59698,2805.26766,1691.15979,243.06505,0.0,0.0,0.0,0.0
+2041,4824.84801,0.0,13.82315,2831.9486,1733.70139,245.37486,0.0,0.0,0.0,0.0
+2042,4897.29233,0.0,14.0536,2858.6044,1776.96353,247.6708,0.0,0.0,0.0,0.0
+2043,4970.42889,0.0,14.28755,2885.24173,1820.94535,249.95426,0.0,0.0,0.0,0.0
+2044,5044.24933,0.0,14.52396,2911.85491,1865.64411,252.22635,0.0,0.0,0.0,0.0
+2045,5118.7412,0.0,14.76191,2938.43567,1911.05584,254.48779,0.0,0.0,0.0,0.0
+2046,5193.90327,0.0,15.00109,2964.98689,1957.17632,256.73897,0.0,0.0,0.0,0.0
+2047,5269.7244,0.0,15.24128,2991.5039,2003.99958,258.97965,0.0,0.0,0.0,0.0
+2048,5346.16451,0.0,15.48174,3017.95737,2051.51664,261.20876,0.0,0.0,0.0,0.0
+2049,5423.17037,0.0,15.72168,3044.30739,2099.71659,263.42471,0.0,0.0,0.0,0.0
+2050,5500.69146,0.0,15.96048,3070.51724,2148.58788,265.62586,0.0,0.0,0.0,0.0
+2051,5578.70692,0.0,16.19784,3096.57572,2198.12176,267.8116,0.0,0.0,0.0,0.0
+2052,5657.18672,0.0,16.43364,3122.46534,2248.30692,269.98082,0.0,0.0,0.0,0.0
+2053,5736.05888,0.0,16.66779,3148.13496,2299.12556,272.13057,0.0,0.0,0.0,0.0
+2054,5815.23808,0.0,16.90025,3173.52337,2350.55718,274.25728,0.0,0.0,0.0,0.0
+2055,5894.65704,0.0,17.13109,3198.58449,2402.58342,276.35803,0.0,0.0,0.0,0.0
+2056,5974.28702,0.0,17.36026,3223.3028,2455.19229,278.43168,0.0,0.0,0.0,0.0
+2057,6054.11935,0.0,17.58787,3247.68046,2508.37334,280.47767,0.0,0.0,0.0,0.0
+2058,6134.13467,0.0,17.8144,3271.71389,2562.11186,282.49452,0.0,0.0,0.0,0.0
+2059,6214.32057,0.0,18.04046,3295.40651,2616.3928,284.4808,0.0,0.0,0.0,0.0
+2060,6294.67271,0.0,18.2666,3318.76742,2671.20309,286.43561,0.0,0.0,0.0,0.0
