+++ solution/health_and_education/data/ref1_population.csv
@@ -1,0 +1,48 @@
+Year,World,OECD90,Eastern Europe,Asia (Sans Japan),Middle East and Africa,Latin America,China,India,EU,USA
+2014,7349.4721,929.27447,407.26543,3957.23614,1421.2991,634.39696,1376.04894,1311.05053,738.44207,321.77363
+2015,7349.4721,929.27447,407.26543,3957.23614,1421.2991,634.39696,1376.04894,1311.05053,738.44207,321.77363
+2016,7436.87553,932.98356,407.76256,3997.60036,1456.99548,641.53357,1382.03581,1328.33174,738.66684,324.11007
+2017,7525.57132,936.64552,408.13764,4038.22818,1493.77952,648.78045,1387.53123,1346.24485,738.76532,326.45235
+2018,7615.15654,940.23746,408.38598,4078.85703,1531.57068,656.10539,1392.45791,1364.65105,738.72382,328.79433
+2019,7705.05663,943.7163,408.51214,4119.13221,1570.23805,663.45793,1396.73688,1383.32427,738.5237,331.12408
+2020,7794.83701,947.05113,408.52045,4158.78543,1609.68202,670.79798,1400.31454,1402.0908,738.15343,333.43229
+2021,7884.32713,950.24158,408.4041,4197.69154,1649.87687,678.11303,1403.15315,1420.90183,737.61047,335.71825
+2022,7973.57998,953.30379,408.15896,4235.87072,1690.84245,685.40406,1405.26334,1439.78526,736.9049,337.98396
+2023,8062.65101,956.24165,407.79768,4273.37667,1732.57583,692.65919,1406.70593,1458.72913,736.05378,340.22541
+2024,8151.67722,959.06165,407.33887,4310.32492,1775.08386,699.86792,1407.5748,1477.73702,735.08089,342.43789
+2025,8240.73751,961.76867,406.79832,4346.78656,1818.36488,707.01908,1407.94727,1496.79674,734.0059,344.61727
+2026,8329.83301,964.364,406.18208,4382.76325,1862.41853,714.10516,1407.84261,1515.89426,732.83581,346.7617
+2027,8418.84153,966.84648,405.49364,4418.17013,1907.21781,721.11347,1407.26714,1534.97087,731.57399,348.86935
+2028,8507.57937,969.21588,404.74371,4452.89309,1952.7044,728.02229,1406.2719,1553.91809,730.23226,350.93621
+2029,8595.80845,971.47142,403.94397,4486.78247,1998.80421,734.80639,1404.91436,1572.60094,728.82308,352.95787
+2030,8683.36509,973.61287,403.10469,4519.7331,2045.46704,741.4474,1403.24039,1590.92081,727.35638,354.93113
+2031,8770.18454,975.64161,402.23198,4551.70732,2092.66837,747.93527,1401.27287,1608.83977,725.83887,356.8543
+2032,8856.33181,977.55957,401.33138,4582.74008,2140.42916,754.27163,1399.02158,1626.3736,724.27351,358.72765
+2033,8941.93799,979.36778,400.4115,4612.89128,2188.80223,760.46519,1396.50051,1643.54794,722.66122,360.55277
+2034,9027.20128,981.06739,399.48118,4642.25569,2237.86573,766.5313,1393.71812,1660.41336,721.00069,362.33245
+2035,9112.2797,982.65999,398.54706,4670.91108,2287.6793,772.48227,1390.67752,1677.01608,719.29074,364.06913
+2036,9197.21086,984.14744,397.61388,4698.87248,2338.25724,778.31983,1387.38911,1693.35827,717.53314,365.76405
+2037,9281.99269,985.53249,396.68256,4726.14354,2389.59096,784.04315,1383.84809,1709.45156,715.72891,367.4179
+2038,9366.69854,986.81887,395.75072,4752.77934,2441.68994,789.65968,1380.02124,1725.36678,713.87455,369.03215
+2039,9451.40531,988.01093,394.81348,4778.8435,2494.55933,795.17807,1375.86196,1741.19345,711.96499,370.60839
+2040,9536.17204,989.11278,393.86679,4804.38497,2548.20211,800.60539,1371.33365,1757.00124,709.99618,372.14854
+2041,9621.03717,990.12662,392.91082,4829.43304,2602.62125,805.94543,1366.42996,1772.81742,707.96744,373.65385
+2042,9706.01075,991.05521,391.94586,4853.99271,2657.81694,811.20003,1361.15238,1788.6447,705.87876,375.12681
+2043,9791.08359,991.90379,390.96795,4878.05675,2713.78353,816.37157,1355.48244,1804.49598,703.72771,376.5721
+2044,9876.22866,992.67835,389.97202,4901.60485,2770.51176,821.46169,1349.40047,1820.37611,701.5115,377.99539
+2045,9961.42218,993.3836,388.95422,4924.62206,2827.99069,826.47161,1342.89522,1836.28576,699.22835,379.40119
+2046,10046.66459,994.02413,387.91326,4947.11302,2886.21171,831.40249,1335.96496,1852.23013,696.87863,380.79271
+2047,10131.95498,994.60116,386.84928,4969.08724,2945.16235,836.25496,1328.62106,1868.20578,694.46344,382.17095
+2048,10217.26309,995.11199,385.76178,4990.53891,3004.82166,841.02875,1320.88389,1884.19046,691.98308,383.5355
+2049,10302.54974,995.55182,384.65072,5011.45973,3065.16449,845.72297,1312.78215,1900.15253,689.4381,384.88481
+2050,10387.78138,995.91818,383.51642,5031.84353,3126.16688,850.33637,1304.34353,1916.06242,686.8301,386.21835
+2051,10472.95119,996.21243,382.35903,5051.69741,3187.81348,854.86884,1295.5838,1931.91201,684.1611,387.53661
+2052,10558.04989,996.43982,381.17931,5071.02454,3250.08738,859.31883,1286.52095,1947.68674,681.43502,388.84231
+2053,10643.03968,996.60677,379.97926,5089.81136,3312.96072,863.68157,1277.18933,1963.34207,678.65802,390.13969
+2054,10727.87591,996.72148,378.76146,5108.03962,3376.40241,867.95094,1267.62865,1978.82401,675.83769,391.43405
+2055,10812.52912,996.79117,377.5284,5125.70173,3440.38571,872.12211,1257.87463,1994.09152,672.98112,392.7294
+2056,10896.99478,996.81983,376.28187,5142.80541,3504.8939,876.19378,1247.94803,2009.13046,670.09318,394.02695
+2057,10981.28851,996.8106,375.02359,5159.37406,3569.91453,880.16573,1237.8687,2023.94178,667.17847,395.32627
+2058,11065.42944,996.76891,373.75611,5175.43721,3635.43145,884.03576,1227.66979,2038.52034,664.24372,396.62742
+2059,11149.44702,996.70044,372.4822,5191.033,3701.4297,887.80168,1217.38672,2052.86665,661.29618,397.92981
+2060,11233.37324,996.61026,371.20461,5206.19887,3767.89712,891.46238,1207.04805,2066.98603,658.34259,399.23256
