+++ solution/health_and_education/data/ref2_low_edu_gender_parity.csv
@@ -1,0 +1,48 @@
+Year,World,OECD90,Eastern Europe,Asia (Sans Japan),Middle East and Africa,Latin America,China,India,EU,USA
+2014,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0
+2015,3104.47059,0.0,8.48186,2065.8178,852.00259,178.16835,0.0,0.0,0.0,0.0
+2016,3157.87084,0.0,8.66946,2092.573,876.10862,180.51975,0.0,0.0,0.0,0.0
+2017,3211.30766,0.0,8.85812,2118.97437,900.63206,182.84313,0.0,0.0,0.0,0.0
+2018,3264.76518,0.0,9.04672,2145.00457,925.57677,185.13711,0.0,0.0,0.0,0.0
+2019,3318.25114,0.0,9.23406,2170.65849,950.95757,187.40103,0.0,0.0,0.0,0.0
+2020,3371.76847,0.0,9.41913,2195.92908,976.78637,189.6339,0.0,0.0,0.0,0.0
+2021,3425.28557,0.0,9.60148,2220.79237,1003.05767,191.83405,0.0,0.0,0.0,0.0
+2022,3478.77262,0.0,9.78092,2245.22595,1029.7664,193.99934,0.0,0.0,0.0,0.0
+2023,3532.24074,0.0,9.95717,2269.22863,1056.92708,196.12787,0.0,0.0,0.0,0.0
+2024,3585.71316,0.0,10.13005,2292.8058,1084.55961,198.2177,0.0,0.0,0.0,0.0
+2025,3639.20392,0.0,10.29952,2315.95853,1112.67872,200.26714,0.0,0.0,0.0,0.0
+2026,3692.70437,0.0,10.46547,2338.6772,1141.28675,202.27495,0.0,0.0,0.0,0.0
+2027,3746.19506,0.0,10.62803,2360.94646,1170.38044,204.24013,0.0,0.0,0.0,0.0
+2028,3799.66619,0.0,10.78779,2382.75504,1199.96176,206.1616,0.0,0.0,0.0,0.0
+2029,3853.10608,0.0,10.94556,2404.09065,1230.03146,208.03841,0.0,0.0,0.0,0.0
+2030,3906.50103,0.0,11.10204,2424.94118,1260.58802,209.86979,0.0,0.0,0.0,0.0
+2031,3959.84065,0.0,11.25742,2445.30062,1291.62773,211.65488,0.0,0.0,0.0,0.0
+2032,4013.10831,0.0,11.41193,2465.15923,1323.14393,213.39321,0.0,0.0,0.0,0.0
+2033,4066.27568,0.0,11.56632,2484.49702,1355.12731,215.08503,0.0,0.0,0.0,0.0
+2034,4119.309,0.0,11.7215,2503.29046,1387.56614,216.7309,0.0,0.0,0.0,0.0
+2035,4172.17921,0.0,11.87814,2521.52214,1420.44772,218.33121,0.0,0.0,0.0,0.0
+2036,4224.8634,0.0,12.03645,2539.1796,1453.76163,219.88573,0.0,0.0,0.0,0.0
+2037,4277.34792,0.0,12.19636,2556.26285,1487.49472,221.39399,0.0,0.0,0.0,0.0
+2038,4329.6238,0.0,12.35787,2572.78277,1521.62736,222.8558,0.0,0.0,0.0,0.0
+2039,4381.68639,0.0,12.52081,2588.75763,1556.13703,224.27092,0.0,0.0,0.0,0.0
+2040,4433.52631,0.0,12.68498,2604.19966,1591.00258,225.6391,0.0,0.0,0.0,0.0
+2041,4485.13066,0.0,12.85032,2619.11179,1626.20838,226.96017,0.0,0.0,0.0,0.0
+2042,4536.47563,0.0,13.01656,2633.48566,1661.73955,228.23386,0.0,0.0,0.0,0.0
+2043,4587.52794,0.0,13.18285,2647.30893,1697.5764,229.45976,0.0,0.0,0.0,0.0
+2044,4638.24828,0.0,13.3481,2660.56442,1733.69837,230.63739,0.0,0.0,0.0,0.0
+2045,4688.60419,0.0,13.51141,2673.24003,1770.08642,231.76633,0.0,0.0,0.0,0.0
+2046,4738.57322,0.0,13.67245,2685.33108,1806.72343,232.84627,0.0,0.0,0.0,0.0
+2047,4788.14492,0.0,13.83102,2696.84193,1843.59496,233.87701,0.0,0.0,0.0,0.0
+2048,4837.31373,0.0,13.98669,2707.78058,1880.68803,234.85843,0.0,0.0,0.0,0.0
+2049,4886.07986,0.0,14.139,2718.15927,1917.99112,235.79048,0.0,0.0,0.0,0.0
+2050,4934.43967,0.0,14.28761,2727.98702,1955.49202,236.67302,0.0,0.0,0.0,0.0
+2051,4982.38243,0.0,14.43231,2737.26795,1993.17616,237.50601,0.0,0.0,0.0,0.0
+2052,5029.89046,0.0,14.57306,2745.99994,2031.02828,238.28918,0.0,0.0,0.0,0.0
+2053,5076.94541,0.0,14.70994,2754.17866,2069.03481,239.022,0.0,0.0,0.0,0.0
+2054,5123.52646,0.0,14.84321,2761.79715,2107.1823,239.7038,0.0,0.0,0.0,0.0
+2055,5169.61579,0.0,14.97308,2768.85153,2145.45707,240.33411,0.0,0.0,0.0,0.0
+2056,5215.20349,0.0,15.09957,2775.34612,2183.84488,240.91292,0.0,0.0,0.0,0.0
+2057,5260.28197,0.0,15.22279,2781.28761,2222.33114,241.44043,0.0,0.0,0.0,0.0
+2058,5304.83924,0.0,15.34313,2786.6781,2260.90125,241.91676,0.0,0.0,0.0,0.0
+2059,5348.86356,0.0,15.46109,2791.5199,2299.54044,242.34214,0.0,0.0,0.0,0.0
+2060,5392.34612,0.0,15.57707,2795.81825,2338.2339,242.71689,0.0,0.0,0.0,0.0
