+++ solution/health_and_education/data/ref2_population.csv
@@ -1,0 +1,48 @@
+Year,World,OECD90,Eastern Europe,Asia (Sans Japan),Middle East and Africa,Latin America,China,India,EU,USA
+2014,7349.4721,929.27447,407.26543,3957.23614,1421.2991,634.39696,1376.04894,1311.05053,738.44207,321.77363
+2015,7349.4721,929.27447,407.26543,3957.23614,1421.2991,634.39696,1376.04894,1311.05053,738.44207,321.77363
+2016,7429.0469,932.21212,407.45046,3993.32073,1455.38354,640.68005,1380.8099,1326.80158,738.13933,323.80581
+2017,7506.4945,934.7676,407.37911,4027.8101,1489.83721,646.70047,1384.55993,1342.51271,737.48253,325.71103
+2018,7581.94776,936.97487,407.07621,4060.75475,1524.66317,652.47877,1387.34872,1358.13772,736.50331,327.50498
+2019,7655.70393,938.88198,406.5918,4092.30245,1559.87564,658.05206,1389.28982,1373.60507,735.25383,329.21096
+2020,7727.98938,940.52724,405.96404,4122.56705,1595.48335,663.4477,1390.47496,1388.85892,733.77575,330.8462
+2021,7798.80624,941.93026,405.19582,4151.52645,1631.48388,668.66984,1390.90742,1403.88269,732.08221,332.41769
+2022,7868.06903,943.0941,404.27486,4179.12802,1667.864,673.70806,1390.57096,1418.68092,730.17184,333.92166
+2023,7935.83328,944.02159,403.20904,4205.43332,1704.61036,678.55897,1389.52427,1433.24267,728.05662,335.3519
+2024,8002.16539,944.71002,402.0067,4230.52591,1741.70657,683.21621,1387.84076,1447.56046,725.74731,336.69791
+2025,8067.13745,945.16269,400.67737,4254.47802,1779.14258,687.6768,1385.58675,1461.62523,723.25708,337.95337
+2026,8130.7737,945.38522,399.2281,4277.30957,1816.91113,691.93967,1382.79074,1475.4246,720.59558,339.11737
+2027,8193.14307,945.39902,397.67033,4299.04193,1855.01963,696.01217,1379.48547,1488.94304,717.78146,340.19725
+2028,8254.41645,945.24169,396.02342,4319.7504,1893.49113,699.90981,1375.74637,1502.1652,714.84753,341.20478
+2029,8314.80408,944.96007,394.3102,4339.5228,1932.358,703.65301,1371.65981,1515.07492,711.83288,342.15572
+2030,8374.46188,944.58933,392.54952,4358.42275,1971.64391,707.25638,1367.2919,1527.65799,708.76781,343.06193
+2031,8433.46137,944.14678,390.75183,4376.4879,2011.34884,710.72602,1362.68155,1539.90705,705.66891,343.92786
+2032,8491.78908,943.63083,388.92263,4393.71195,2051.4654,714.05827,1357.83484,1551.81449,702.53891,344.75221
+2033,8549.40555,943.03358,387.06925,4410.05739,2091.99969,717.24564,1352.7414,1563.36648,699.37626,345.53383
+2034,8606.23002,942.33805,385.19729,4425.46137,2132.95742,720.2759,1347.37308,1574.54824,696.17264,346.26933
+2035,8662.19575,941.53143,383.31137,4439.87582,2174.33746,723.13968,1341.70792,1585.34985,692.922,346.95606
+2036,8717.29157,940.61393,381.41603,4453.29188,2216.13458,725.83516,1335.75157,1595.76373,689.62606,347.59488
+2037,8771.51987,939.59072,379.51468,4465.71963,2258.33081,728.36403,1329.51323,1605.7925,686.28973,348.18691
+2038,8824.83704,938.45902,377.60858,4477.15076,2300.8944,730.72429,1322.97653,1615.44985,682.91268,348.7297
+2039,8877.19081,937.21549,375.69756,4487.57885,2343.78464,732.91426,1316.11947,1624.75632,679.49396,349.21999
+2040,8928.53173,935.85684,373.78066,4496.99717,2386.96465,734.93242,1308.92527,1633.72767,676.03149,349.65532
+2041,8978.83217,934.38097,371.85874,4505.40248,2430.41227,736.77771,1301.38902,1642.36934,672.52447,350.03486
+2042,9028.05926,932.78524,369.93007,4512.78696,2474.10809,738.4489,1293.51221,1650.67962,668.96892,350.35901
+2043,9076.15169,931.06514,367.98721,4519.13481,2518.02014,739.94438,1285.2926,1658.65883,665.35528,350.62807
+2044,9123.03978,929.21549,366.02024,4524.42735,2562.11423,741.26247,1276.72972,1666.30529,661.67138,350.84278
+2045,9168.66694,927.23193,364.02109,4528.65275,2606.35895,742.40221,1267.82422,1673.61877,657.90737,351.00378
+2046,9213.00344,925.115,361.98722,4531.80876,2650.72927,743.36319,1258.57761,1680.59958,654.06081,351.11249
+2047,9256.03432,922.86374,359.91768,4533.90353,2695.20322,744.14616,1248.99236,1687.25222,650.13048,351.16917
+2048,9297.73877,920.47113,357.80801,4534.94992,2739.75668,744.75304,1239.07241,1693.58529,646.10954,351.17223
+2049,9338.10174,917.92853,355.65362,4534.96658,2784.36647,745.18655,1228.82293,1699.60963,641.99059,351.11934
+2050,9377.11399,915.23136,353.45131,4533.97137,2829.01083,745.44912,1218.25063,1705.33254,637.76923,351.00981
+2051,9414.77452,912.37976,351.19962,4531.98379,2873.66898,745.54238,1207.36871,1710.75814,633.44478,350.84386
+2052,9451.0863,909.38082,348.89906,4529.01687,2918.32212,745.46743,1196.18811,1715.88335,629.02114,350.62519
+2053,9486.05373,906.24636,346.55133,4525.07653,2962.95366,745.22584,1184.71221,1720.69922,624.50499,350.36097
+2054,9519.6855,902.99235,344.15935,4520.16601,3007.54865,744.81914,1172.94273,1725.19268,619.90583,350.06054
+2055,9551.99895,899.63305,341.72617,4514.29804,3052.09257,744.24913,1160.88878,1729.35406,615.23296,349.73145
+2056,9583.00809,896.17543,339.25362,4507.4917,3096.56941,743.51792,1148.56052,1733.18387,610.49187,349.37648
+2057,9612.74763,892.62584,336.74406,4499.78531,3140.96437,742.62805,1135.98562,1736.68515,605.68893,348.99659
+2058,9641.28447,888.99776,334.20199,4491.23616,3185.26616,741.5824,1123.21312,1739.85419,600.83602,348.59511
+2059,9668.70217,885.30621,331.63285,4481.91392,3229.46502,740.38418,1110.30397,1742.6869,595.94692,348.17518
+2060,9695.07315,881.56408,329.0418,4471.88005,3273.55069,739.03653,1097.30865,1745.18241,591.03419,347.73901
