--- solution/healthandeducation/__init__.py
@@ -1,460 +1,0 @@
-"""Health & Education solution model for Electricity Cluster
-   Excel filename: CORE_PopulationChange_29Jan2020 (version 1.4).xlsx
-   Excel sheet name: Electricity_cluster
-"""
-
-import pathlib
-
-import numpy as np
-import pandas as pd
-
-# import sys
-# sys.path.append('c:\\Users\\sunishchal.dev\\Documents\\solutions')
-
-# import solarpvutil
-
-DATADIR = pathlib.Path(__file__).parents[2].joinpath('data')
-THISDIR = pathlib.Path(__file__).parents[0]
-
-name = 'Health and Education - Electricity Cluster'
-# solution_category = ac.SOLUTION_CATEGORY.REDUCTION #TODO: Confirm this is a reduction solution
-
-# Assumptions:
-# % impact of educational attainment on uptake of Family Planning:
-fixed_weighting_factor = None
-pct_impact = 0.50
-use_fixed_weight = 'N'
-
-# Regions included as LLDC+HighNRR:
-lldc_high_nrr_config = {
-    'OECD90': 'N', 
-    'Eastern Europe': 'N', 
-    'Asia (Sans Japan)': 'Y', 
-    'Middle East and Africa': 'Y', 
-    'Latin America': 'N'
-}
-lldc_high_nrr_regions_y = dict(filter(lambda x: x[1] == 'Y', lldc_high_nrr_config.items())).keys()
-lldc_high_nrr_regions_n = dict(filter(lambda x: x[1] == 'N', lldc_high_nrr_config.items())).keys()
-
-
-class Scenario:
-    name = name
-    # solution_category = solution_category
-
-    def __init__(self, scenario=None):
-
-        # Population scenarios
-        # Population_Tables!C2:L49
-        self.ref1_population = pd.DataFrame(ref1_population_list[1:],
-            columns=ref1_population_list[0],
-            index=list(range(2014, 2061)), dtype=np.float64)
-
-        # Population_Tables!O2:X49
-        self.ref2_population = pd.DataFrame(ref2_population_list[1:],
-            columns=ref2_population_list[0],
-            index=list(range(2014, 2061)), dtype=np.float64)
-        
-        # Population_Tables!C61:L108
-        self.ref1_low_edu = pd.DataFrame(ref1_low_edu_list[1:],
-            columns=ref1_low_edu_list[0],
-            index=list(range(2014, 2061)), dtype=np.float64)
-
-        # Population_Tables!O61:X108
-        self.ref2_low_edu = pd.DataFrame(ref2_low_edu_list[1:],
-            columns=ref2_low_edu_list[0],
-            index=list(range(2014, 2061)), dtype=np.float64)
-
-        # TABLE 1: Current TAM Mix
-        current_tam_mix = pd.DataFrame(current_tam_mix_list,
-                    columns=['2018', 'Include in SOL?', 'Include in CONV?'],
-                    index=['Coal',
-                            'Natural gas',
-                            'Nuclear',
-                            'Oil',
-                            'Hydroelectric',
-                            'Solar Photovoltaic',
-                            'Wave and Tidal',
-                            'Wind Onshore',
-                            'Wind Offshore',
-                            'Biomass and Waste',
-                            'Concentrated Solar Power',
-                            'Geothermal'])
-
-        # Table 2: REF2, Electricity Generation TAM (TWh)										 
-        # Electricity_cluster!B26:K73
-        # obj = solarpvutil.Scenario() # Not using electricity TAM output due to slight discrepancy in Asia values
-        # self.ref2_tam = obj.tm.ref_tam_per_region()
-        self.ref2_tam = pd.DataFrame(ref2_tam_list[1:],
-            columns=ref2_tam_list[0],
-            index=list(range(2014, 2061)), dtype=np.float64)
-            
-        # Table 3: REF1 ,Electricity Generation TAM (TWh)	
-        # (a) FOR POPULATIONS IN REGIONS WITH LOW EDUCATIONAL ATTAINMENT ONLY!	
-        ref1_tam_low_edu = (self.ref2_tam / self.ref2_population) * self.ref1_low_edu
-        ref1_tam_low_edu.loc[:, 'Asia (Sans Japan)'] = ((self.ref2_tam.loc[:, 'Asia (Sans Japan)'] - self.ref2_tam.loc[:, 'China']) / self.ref2_population.loc[:, 'Asia (Sans Japan)']) * self.ref1_low_edu.loc[:, 'Asia (Sans Japan)']
-        ref1_tam_low_edu.loc[:, 'World'] = ref1_tam_low_edu.loc[:, ref1_tam_low_edu.columns[1:6]].sum(axis=1)									
-								
-        # Electricity_cluster!M26:W73
-        self.ref1_tam_low_edu = ref1_tam_low_edu
-
-        # Table 3: REF1 ,Electricity Generation TAM (TWh)
-        # (b) FOR REGIONS WITH HIGHER EDUCATIONAL ATTAINMENT										
-        ref1_tam_high_edu = ((self.ref2_tam / self.ref2_population) * self.ref1_population) - ref1_tam_low_edu
-
-        # Electricity_cluster!@26:AI73
-        self.ref1_tam_high_edu = ref1_tam_high_edu
-
-        # Table 3: REF1 ,Electricity Generation TAM (TWh)	
-        # (c) FOR ALL REGIONS										
-        ref1_tam_all_regions = ref1_tam_low_edu + ref1_tam_high_edu
-
-        # Electricity_cluster!AL26:AU73
-        self.ref1_tam_all_regions = ref1_tam_all_regions
-
-        # Table 4: Total REF2 Electricity Generation by Economic Development Status (TWh)					
-        ref2_elec_gen = pd.DataFrame(None,
-                    columns=['LLDC+HighNRR', 'China', 'MDC + LAC + EE', 'Total Electricity Demand in Countries with Higher Educational Attainment (TWh)', 'Total Electricity Demand in Countries with Higher Educational Attainment (TWh) % LLDC'],
-                    index=list(range(2014, 2061)), dtype=np.float64)
-
-        ref2_elec_gen.loc[:, 'LLDC+HighNRR'] = self.ref2_tam.loc[:, lldc_high_nrr_regions_y].sum(axis=1) - self.ref2_tam.loc[:, 'China']
-        ref2_elec_gen.loc[:, 'China'] = self.ref2_tam.loc[:, 'China']
-        ref2_elec_gen.loc[:, 'MDC + LAC + EE'] = self.ref2_tam.loc[:, lldc_high_nrr_regions_n].sum(axis=1) 
-        if lldc_high_nrr_config['Asia (Sans Japan)'] == 'N':
-            ref2_elec_gen.loc[:, 'MDC + LAC + EE'] = ref2_elec_gen.loc[:, 'MDC + LAC + EE'] - self.ref2_tam.loc[:, 'China']
-        ref2_elec_gen.loc[:, 'Total Electricity Demand in Countries with Higher Educational Attainment (TWh)'] = ref2_elec_gen.loc[:, ['LLDC+HighNRR', 'China', 'MDC + LAC + EE']].sum(axis=1)
-        ref2_elec_gen.loc[:, 'Total Electricity Demand in Countries with Higher Educational Attainment (TWh) % LLDC'] = ref2_elec_gen.loc[:, 'LLDC+HighNRR'] / ref2_elec_gen.loc[:, 'Total Electricity Demand in Countries with Higher Educational Attainment (TWh)']
-
-        # Electricity_cluster!B77:F124
-        self.ref2_elec_gen = ref2_elec_gen
-
-        # Table 5: Total REF1 Electricity Generation by Economic Development Status (TWh)												
-        ref1_elec_gen = pd.DataFrame(None,
-                    columns=['LLDC with low educational attainment, excluding China', 'MDC + EE + LAC with low educational attainment, excluding China', 'China', 'LLDC with higher educational attainment, excluding China', 'MDC + EE + LAC with higher educational attainment', 'Total Electricity Demand in Countries with Low Educational Attainment (TWh), exluding China', 'Total Electricity Demand in Countries with Low Educational Attainment (TWh), exluding China % LLDC', 'Total Electricity Demand in Countries with Higher Educational Attainment (TWh)', 'Total Electricity Demand in Countries with Higher Educational Attainment (TWh) % LLDC', 'Total Electricity Demand (TWh)', 'Total Electricity Demand (TWh) % LLDC'],
-                    index=list(range(2014, 2061)), dtype=np.float64)
-
-        ref1_elec_gen.loc[:, 'LLDC with low educational attainment, excluding China'] = ref1_tam_low_edu.loc[:, lldc_high_nrr_regions_y].sum(axis=1) - ref1_tam_low_edu.loc[:, 'China']
-        ref1_elec_gen.loc[:, 'MDC + EE + LAC with low educational attainment, excluding China'] = ref1_tam_low_edu.loc[:, lldc_high_nrr_regions_n].sum(axis=1)
-        if lldc_high_nrr_config['Asia (Sans Japan)'] == 'N':
-            ref1_elec_gen.loc[:, 'MDC + LAC + EE'] = ref1_elec_gen.loc[:, 'MDC + LAC + EE'] - self.ref1_tam_low_edu.loc[:, 'China']
-        ref1_elec_gen.loc[:, 'China'] = ref1_tam_high_edu.loc[:, 'China']
-
-        ref1_elec_gen.loc[:, 'LLDC with higher educational attainment, excluding China'] = ref1_tam_high_edu.loc[:, lldc_high_nrr_regions_y].sum(axis=1) - ref1_tam_high_edu.loc[:, 'China']
-        ref1_elec_gen.loc[:, 'MDC + EE + LAC with higher educational attainment'] = ref1_tam_high_edu.loc[:, lldc_high_nrr_regions_n].sum(axis=1)
-        if lldc_high_nrr_config['Asia (Sans Japan)'] == 'N':
-            ref1_elec_gen.loc[:, 'MDC + EE + LAC with higher educational attainment'] = ref1_elec_gen.loc[:, 'MDC + EE + LAC with higher educational attainment'] - self.ref1_tam_high_edu.loc[:, 'China']
-
-        ref1_elec_gen.loc[:, 'Total Electricity Demand in Countries with Low Educational Attainment (TWh), exluding China'] = ref1_elec_gen.loc[:, 'LLDC with low educational attainment, excluding China'] + ref1_elec_gen.loc[:, 'MDC + EE + LAC with low educational attainment, excluding China']
-        ref1_elec_gen.loc[:, 'Total Electricity Demand in Countries with Low Educational Attainment (TWh), exluding China % LLDC'] = ref1_elec_gen.loc[:, 'LLDC with low educational attainment, excluding China'] / ref1_elec_gen.loc[:, 'Total Electricity Demand in Countries with Low Educational Attainment (TWh), exluding China']
-
-        ref1_elec_gen.loc[:, 'Total Electricity Demand in Countries with Higher Educational Attainment (TWh)'] = ref1_elec_gen.loc[:, 'China'] + ref1_elec_gen.loc[:, 'LLDC with higher educational attainment, excluding China'] + ref1_elec_gen.loc[:, 'MDC + EE + LAC with higher educational attainment']
-        ref1_elec_gen.loc[:, 'Total Electricity Demand in Countries with Higher Educational Attainment (TWh) % LLDC'] = ref1_elec_gen.loc[:, 'LLDC with higher educational attainment, excluding China'] / ref1_elec_gen.loc[:, 'Total Electricity Demand in Countries with Higher Educational Attainment (TWh)']
-
-        ref1_elec_gen.loc[:, 'Total Electricity Demand (TWh)'] = ref1_elec_gen.loc[:, 'Total Electricity Demand in Countries with Low Educational Attainment (TWh), exluding China'] + ref1_elec_gen.loc[:, 'Total Electricity Demand in Countries with Higher Educational Attainment (TWh)']
-        ref1_elec_gen.loc[:, 'Total Electricity Demand (TWh) % LLDC'] = ref1_elec_gen.loc[:, 'Total Electricity Demand in Countries with Low Educational Attainment (TWh), exluding China'] / ref1_elec_gen.loc[:, 'Total Electricity Demand (TWh)']
-
-        # Electricity_cluster!H77:T124
-        self.ref1_elec_gen = ref1_elec_gen
-
-        # Table 6: Change in Electricity Generation by MDC vs. LLDC Regions, REF1-REF2 (TWh)							
-        change_elec_gen = pd.DataFrame(None,
-                    columns=['LLDC', 'China', 'MDC + EE +LAC', 'Total change in REF1-REF2', '% LLDC with higher educational attainment', '% LLDC with Low Educational Attainment', '% LLDC', '% MDC + LAC + EE + China'],
-                    index=list(range(2014, 2061)), dtype=np.float64)
-
-        change_elec_gen.loc[:, 'LLDC'] = (ref1_elec_gen.loc[:, 'LLDC with low educational attainment, excluding China'] + ref1_elec_gen.loc[:, 'LLDC with higher educational attainment, excluding China']) - ref2_elec_gen.loc[:, 'LLDC+HighNRR']
-        change_elec_gen.loc[:, 'China'] = ref1_elec_gen.loc[:, 'China'] - ref2_elec_gen.loc[:, 'China']
-        change_elec_gen.loc[:, 'MDC + EE +LAC'] = (ref1_elec_gen.loc[:, 'MDC + EE + LAC with higher educational attainment'] + ref1_elec_gen.loc[:, 'MDC + EE + LAC with low educational attainment, excluding China']) - ref2_elec_gen.loc[:, 'MDC + LAC + EE']
-        change_elec_gen.loc[[2014, 2015], 'MDC + EE +LAC'] = 0
-
-        change_elec_gen.loc[:, 'Total change in REF1-REF2'] = change_elec_gen.loc[:, 'LLDC'] + change_elec_gen.loc[:, 'China'] + change_elec_gen.loc[:, 'MDC + EE +LAC']
-
-        change_elec_gen.loc[:, '% LLDC with higher educational attainment'] = (change_elec_gen.loc[:, 'LLDC'] * (ref1_elec_gen.loc[:, 'LLDC with higher educational attainment, excluding China'] / (ref1_elec_gen.loc[:, 'LLDC with low educational attainment, excluding China'] + ref1_elec_gen.loc[:, 'LLDC with higher educational attainment, excluding China']))) / change_elec_gen.loc[:, 'Total change in REF1-REF2']
-
-        change_elec_gen.loc[:, '% LLDC with Low Educational Attainment'] = (change_elec_gen.loc[:, 'LLDC'] / change_elec_gen.loc[:, 'Total change in REF1-REF2']) * (ref1_elec_gen.loc[:, 'LLDC with low educational attainment, excluding China'] / (ref1_elec_gen.loc[:, 'LLDC with higher educational attainment, excluding China'] + ref1_elec_gen.loc[:, 'LLDC with low educational attainment, excluding China'])) 
-        change_elec_gen.loc[:, '% LLDC'] = (change_elec_gen.loc[:, 'LLDC'] / change_elec_gen.loc[:, 'Total change in REF1-REF2'])
-        change_elec_gen.loc[:, '% MDC + LAC + EE + China'] = (change_elec_gen.loc[:, 'China'] + change_elec_gen.loc[:, 'MDC + EE +LAC']) / change_elec_gen.loc[:, 'Total change in REF1-REF2']
-
-        # Electricity_cluster!W77:AD124
-        self.change_elec_gen = change_elec_gen
-
-        # Table 7: Difference in FUNCTIONAL & IMPLEMENTATION UNITS between REF1 and REF2 populations in LLDC+HighNRR+HighED	
-        # CONVENTIONAL												
-        addl_func_units_highed = pd.DataFrame(None,
-                    columns=['Additional Functional Units in REF2 vs REF2 (TWh)', 'Annual Functional Units Increase (TWh)', 'Change in TAM (%)', 'Annual Implementation Units Increase + Replacement (TW)'],
-                    index=list(range(2014, 2061)), dtype=np.float64)
-        
-
-# TABLE 1: Current TAM Mix
-current_tam_mix_list = [
-        [39.28, 'N', 'Y'],
-        [22.72, 'N', 'Y'],
-        [10.45, 'N', 'N'],
-        [3.41, 'N', 'Y'],
-        [15.52, 'N', 'Y'],
-        [1.73, 'N', 'N'],
-        [0.00, 'N', 'N'],
-        [4.36, 'N', 'N'],
-        [0.24, 'N', 'N'],
-        [1.90, 'N', 'N'],
-        [0.05, 'N', 'N'],
-        [0.34, 'N', 'N']]
-
-# REF1 Population
-# Population_Tables!C2:L49
-ref1_population_list = [
-        ['World', 'OECD90', 'Eastern Europe', 'Asia (Sans Japan)', 'Middle East and Africa', 'Latin America', 'China', 'India', 'EU', 'USA'],
-        [7349.47210, 929.27447, 407.26543, 3957.23614, 1421.29910, 634.39696, 1376.04894, 1311.05053, 738.44207, 321.77363],
-        [7349.47210, 929.27447, 407.26543, 3957.23614, 1421.29910, 634.39696, 1376.04894, 1311.05053, 738.44207, 321.77363],
-        [7436.87553, 932.98356, 407.76256, 3997.60036, 1456.99548, 641.53357, 1382.03581, 1328.33174, 738.66684, 324.11007],
-        [7525.57132, 936.64552, 408.13764, 4038.22818, 1493.77952, 648.78045, 1387.53123, 1346.24485, 738.76532, 326.45235],
-        [7615.15654, 940.23746, 408.38598, 4078.85703, 1531.57068, 656.10539, 1392.45791, 1364.65105, 738.72382, 328.79433],
-        [7705.05663, 943.71630, 408.51214, 4119.13221, 1570.23805, 663.45793, 1396.73688, 1383.32427, 738.52370, 331.12408],
-        [7794.83701, 947.05113, 408.52045, 4158.78543, 1609.68202, 670.79798, 1400.31454, 1402.09080, 738.15343, 333.43229],
-        [7884.32713, 950.24158, 408.40410, 4197.69154, 1649.87687, 678.11303, 1403.15315, 1420.90183, 737.61047, 335.71825],
-        [7973.57998, 953.30379, 408.15896, 4235.87072, 1690.84245, 685.40406, 1405.26334, 1439.78526, 736.90490, 337.98396],
-        [8062.65101, 956.24165, 407.79768, 4273.37667, 1732.57583, 692.65919, 1406.70593, 1458.72913, 736.05378, 340.22541],
-        [8151.67722, 959.06165, 407.33887, 4310.32492, 1775.08386, 699.86792, 1407.57480, 1477.73702, 735.08089, 342.43789],
-        [8240.73751, 961.76867, 406.79832, 4346.78656, 1818.36488, 707.01908, 1407.94727, 1496.79674, 734.00590, 344.61727],
-        [8329.83301, 964.36400, 406.18208, 4382.76325, 1862.41853, 714.10516, 1407.84261, 1515.89426, 732.83581, 346.76170],
-        [8418.84153, 966.84648, 405.49364, 4418.17013, 1907.21781, 721.11347, 1407.26714, 1534.97087, 731.57399, 348.86935],
-        [8507.57937, 969.21588, 404.74371, 4452.89309, 1952.70440, 728.02229, 1406.27190, 1553.91809, 730.23226, 350.93621],
-        [8595.80845, 971.47142, 403.94397, 4486.78247, 1998.80421, 734.80639, 1404.91436, 1572.60094, 728.82308, 352.95787],
-        [8683.36509, 973.61287, 403.10469, 4519.73310, 2045.46704, 741.44740, 1403.24039, 1590.92081, 727.35638, 354.93113],
-        [8770.18454, 975.64161, 402.23198, 4551.70732, 2092.66837, 747.93527, 1401.27287, 1608.83977, 725.83887, 356.85430],
-        [8856.33181, 977.55957, 401.33138, 4582.74008, 2140.42916, 754.27163, 1399.02158, 1626.37360, 724.27351, 358.72765],
-        [8941.93799, 979.36778, 400.41150, 4612.89128, 2188.80223, 760.46519, 1396.50051, 1643.54794, 722.66122, 360.55277],
-        [9027.20128, 981.06739, 399.48118, 4642.25569, 2237.86573, 766.53130, 1393.71812, 1660.41336, 721.00069, 362.33245],
-        [9112.27970, 982.65999, 398.54706, 4670.91108, 2287.67930, 772.48227, 1390.67752, 1677.01608, 719.29074, 364.06913],
-        [9197.21086, 984.14744, 397.61388, 4698.87248, 2338.25724, 778.31983, 1387.38911, 1693.35827, 717.53314, 365.76405],
-        [9281.99269, 985.53249, 396.68256, 4726.14354, 2389.59096, 784.04315, 1383.84809, 1709.45156, 715.72891, 367.41790],
-        [9366.69854, 986.81887, 395.75072, 4752.77934, 2441.68994, 789.65968, 1380.02124, 1725.36678, 713.87455, 369.03215],
-        [9451.40531, 988.01093, 394.81348, 4778.84350, 2494.55933, 795.17807, 1375.86196, 1741.19345, 711.96499, 370.60839],
-        [9536.17204, 989.11278, 393.86679, 4804.38497, 2548.20211, 800.60539, 1371.33365, 1757.00124, 709.99618, 372.14854],
-        [9621.03717, 990.12662, 392.91082, 4829.43304, 2602.62125, 805.94543, 1366.42996, 1772.81742, 707.96744, 373.65385],
-        [9706.01075, 991.05521, 391.94586, 4853.99271, 2657.81694, 811.20003, 1361.15238, 1788.64470, 705.87876, 375.12681],
-        [9791.08359, 991.90379, 390.96795, 4878.05675, 2713.78353, 816.37157, 1355.48244, 1804.49598, 703.72771, 376.57210],
-        [9876.22866, 992.67835, 389.97202, 4901.60485, 2770.51176, 821.46169, 1349.40047, 1820.37611, 701.51150, 377.99539],
-        [9961.42218, 993.38360, 388.95422, 4924.62206, 2827.99069, 826.47161, 1342.89522, 1836.28576, 699.22835, 379.40119],
-        [10046.66459, 994.02413, 387.91326, 4947.11302, 2886.21171, 831.40249, 1335.96496, 1852.23013, 696.87863, 380.79271],
-        [10131.95498, 994.60116, 386.84928, 4969.08724, 2945.16235, 836.25496, 1328.62106, 1868.20578, 694.46344, 382.17095],
-        [10217.26309, 995.11199, 385.76178, 4990.53891, 3004.82166, 841.02875, 1320.88389, 1884.19046, 691.98308, 383.53550],
-        [10302.54974, 995.55182, 384.65072, 5011.45973, 3065.16449, 845.72297, 1312.78215, 1900.15253, 689.43810, 384.88481],
-        [10387.78138, 995.91818, 383.51642, 5031.84353, 3126.16688, 850.33637, 1304.34353, 1916.06242, 686.83010, 386.21835],
-        [10472.95119, 996.21243, 382.35903, 5051.69741, 3187.81348, 854.86884, 1295.58380, 1931.91201, 684.16110, 387.53661],
-        [10558.04989, 996.43982, 381.17931, 5071.02454, 3250.08738, 859.31883, 1286.52095, 1947.68674, 681.43502, 388.84231],
-        [10643.03968, 996.60677, 379.97926, 5089.81136, 3312.96072, 863.68157, 1277.18933, 1963.34207, 678.65802, 390.13969],
-        [10727.87591, 996.72148, 378.76146, 5108.03962, 3376.40241, 867.95094, 1267.62865, 1978.82401, 675.83769, 391.43405],
-        [10812.52912, 996.79117, 377.52840, 5125.70173, 3440.38571, 872.12211, 1257.87463, 1994.09152, 672.98112, 392.72940],
-        [10896.99478, 996.81983, 376.28187, 5142.80541, 3504.89390, 876.19378, 1247.94803, 2009.13046, 670.09318, 394.02695],
-        [10981.28851, 996.81060, 375.02359, 5159.37406, 3569.91453, 880.16573, 1237.86870, 2023.94178, 667.17847, 395.32627],
-        [11065.42944, 996.76891, 373.75611, 5175.43721, 3635.43145, 884.03576, 1227.66979, 2038.52034, 664.24372, 396.62742],
-        [11149.44702, 996.70044, 372.48220, 5191.03300, 3701.42970, 887.80168, 1217.38672, 2052.86665, 661.29618, 397.92981],
-        [11233.37324, 996.61026, 371.20461, 5206.19887, 3767.89712, 891.46238, 1207.04805, 2066.98603, 658.34259, 399.23256]]
-
-# REF 2 Population Table
-# Population_Tables!O2:X49
-ref2_population_list = [
-        ['World', 'OECD90', 'Eastern Europe', 'Asia (Sans Japan)', 'Middle East and Africa', 'Latin America', 'China', 'India', 'EU', 'USA'],
-        [7349.47210, 929.27447, 407.26543, 3957.23614, 1421.29910, 634.39696, 1376.04894, 1311.05053, 738.44207, 321.77363],
-        [7349.47210, 929.27447, 407.26543, 3957.23614, 1421.29910, 634.39696, 1376.04894, 1311.05053, 738.44207, 321.77363],
-        [7429.04690, 932.21212, 407.45046, 3993.32073, 1455.38354, 640.68005, 1380.80990, 1326.80158, 738.13933, 323.80581],
-        [7506.49450, 934.76760, 407.37911, 4027.81010, 1489.83721, 646.70047, 1384.55993, 1342.51271, 737.48253, 325.71103],
-        [7581.94776, 936.97487, 407.07621, 4060.75475, 1524.66317, 652.47877, 1387.34872, 1358.13772, 736.50331, 327.50498],
-        [7655.70393, 938.88198, 406.59180, 4092.30245, 1559.87564, 658.05206, 1389.28982, 1373.60507, 735.25383, 329.21096],
-        [7727.98938, 940.52724, 405.96404, 4122.56705, 1595.48335, 663.44770, 1390.47496, 1388.85892, 733.77575, 330.84620],
-        [7798.80624, 941.93026, 405.19582, 4151.52645, 1631.48388, 668.66984, 1390.90742, 1403.88269, 732.08221, 332.41769],
-        [7868.06903, 943.09410, 404.27486, 4179.12802, 1667.86400, 673.70806, 1390.57096, 1418.68092, 730.17184, 333.92166],
-        [7935.83328, 944.02159, 403.20904, 4205.43332, 1704.61036, 678.55897, 1389.52427, 1433.24267, 728.05662, 335.35190],
-        [8002.16539, 944.71002, 402.00670, 4230.52591, 1741.70657, 683.21621, 1387.84076, 1447.56046, 725.74731, 336.69791],
-        [8067.13745, 945.16269, 400.67737, 4254.47802, 1779.14258, 687.67680, 1385.58675, 1461.62523, 723.25708, 337.95337],
-        [8130.77370, 945.38522, 399.22810, 4277.30957, 1816.91113, 691.93967, 1382.79074, 1475.42460, 720.59558, 339.11737],
-        [8193.14307, 945.39902, 397.67033, 4299.04193, 1855.01963, 696.01217, 1379.48547, 1488.94304, 717.78146, 340.19725],
-        [8254.41645, 945.24169, 396.02342, 4319.75040, 1893.49113, 699.90981, 1375.74637, 1502.16520, 714.84753, 341.20478],
-        [8314.80408, 944.96007, 394.31020, 4339.52280, 1932.35800, 703.65301, 1371.65981, 1515.07492, 711.83288, 342.15572],
-        [8374.46188, 944.58933, 392.54952, 4358.42275, 1971.64391, 707.25638, 1367.29190, 1527.65799, 708.76781, 343.06193],
-        [8433.46137, 944.14678, 390.75183, 4376.48790, 2011.34884, 710.72602, 1362.68155, 1539.90705, 705.66891, 343.92786],
-        [8491.78908, 943.63083, 388.92263, 4393.71195, 2051.46540, 714.05827, 1357.83484, 1551.81449, 702.53891, 344.75221],
-        [8549.40555, 943.03358, 387.06925, 4410.05739, 2091.99969, 717.24564, 1352.74140, 1563.36648, 699.37626, 345.53383],
-        [8606.23002, 942.33805, 385.19729, 4425.46137, 2132.95742, 720.27590, 1347.37308, 1574.54824, 696.17264, 346.26933],
-        [8662.19575, 941.53143, 383.31137, 4439.87582, 2174.33746, 723.13968, 1341.70792, 1585.34985, 692.92200, 346.95606],
-        [8717.29157, 940.61393, 381.41603, 4453.29188, 2216.13458, 725.83516, 1335.75157, 1595.76373, 689.62606, 347.59488],
-        [8771.51987, 939.59072, 379.51468, 4465.71963, 2258.33081, 728.36403, 1329.51323, 1605.79250, 686.28973, 348.18691],
-        [8824.83704, 938.45902, 377.60858, 4477.15076, 2300.89440, 730.72429, 1322.97653, 1615.44985, 682.91268, 348.72970],
-        [8877.19081, 937.21549, 375.69756, 4487.57885, 2343.78464, 732.91426, 1316.11947, 1624.75632, 679.49396, 349.21999],
-        [8928.53173, 935.85684, 373.78066, 4496.99717, 2386.96465, 734.93242, 1308.92527, 1633.72767, 676.03149, 349.65532],
-        [8978.83217, 934.38097, 371.85874, 4505.40248, 2430.41227, 736.77771, 1301.38902, 1642.36934, 672.52447, 350.03486],
-        [9028.05926, 932.78524, 369.93007, 4512.78696, 2474.10809, 738.44890, 1293.51221, 1650.67962, 668.96892, 350.35901],
-        [9076.15169, 931.06514, 367.98721, 4519.13481, 2518.02014, 739.94438, 1285.29260, 1658.65883, 665.35528, 350.62807],
-        [9123.03978, 929.21549, 366.02024, 4524.42735, 2562.11423, 741.26247, 1276.72972, 1666.30529, 661.67138, 350.84278],
-        [9168.66694, 927.23193, 364.02109, 4528.65275, 2606.35895, 742.40221, 1267.82422, 1673.61877, 657.90737, 351.00378],
-        [9213.00344, 925.11500, 361.98722, 4531.80876, 2650.72927, 743.36319, 1258.57761, 1680.59958, 654.06081, 351.11249],
-        [9256.03432, 922.86374, 359.91768, 4533.90353, 2695.20322, 744.14616, 1248.99236, 1687.25222, 650.13048, 351.16917],
-        [9297.73877, 920.47113, 357.80801, 4534.94992, 2739.75668, 744.75304, 1239.07241, 1693.58529, 646.10954, 351.17223],
-        [9338.10174, 917.92853, 355.65362, 4534.96658, 2784.36647, 745.18655, 1228.82293, 1699.60963, 641.99059, 351.11934],
-        [9377.11399, 915.23136, 353.45131, 4533.97137, 2829.01083, 745.44912, 1218.25063, 1705.33254, 637.76923, 351.00981],
-        [9414.77452, 912.37976, 351.19962, 4531.98379, 2873.66898, 745.54238, 1207.36871, 1710.75814, 633.44478, 350.84386],
-        [9451.08630, 909.38082, 348.89906, 4529.01687, 2918.32212, 745.46743, 1196.18811, 1715.88335, 629.02114, 350.62519],
-        [9486.05373, 906.24636, 346.55133, 4525.07653, 2962.95366, 745.22584, 1184.71221, 1720.69922, 624.50499, 350.36097],
-        [9519.68550, 902.99235, 344.15935, 4520.16601, 3007.54865, 744.81914, 1172.94273, 1725.19268, 619.90583, 350.06054],
-        [9551.99895, 899.63305, 341.72617, 4514.29804, 3052.09257, 744.24913, 1160.88878, 1729.35406, 615.23296, 349.73145],
-        [9583.00809, 896.17543, 339.25362, 4507.49170, 3096.56941, 743.51792, 1148.56052, 1733.18387, 610.49187, 349.37648],
-        [9612.74763, 892.62584, 336.74406, 4499.78531, 3140.96437, 742.62805, 1135.98562, 1736.68515, 605.68893, 348.99659],
-        [9641.28447, 888.99776, 334.20199, 4491.23616, 3185.26616, 741.58240, 1123.21312, 1739.85419, 600.83602, 348.59511],
-        [9668.70217, 885.30621, 331.63285, 4481.91392, 3229.46502, 740.38418, 1110.30397, 1742.68690, 595.94692, 348.17518],
-        [9695.07315, 881.56408, 329.04180, 4471.88005, 3273.55069, 739.03653, 1097.30865, 1745.18241, 591.03419, 347.73901]]
-
-# Population in Countries with <0.96 gender partiy in completing upper secondary education, by IPCC Region, REF1_population_scenario (millions)
-# Population_Tables!C61:L108
-ref1_low_edu_list = [
-        ['World', 'OECD90', 'Eastern Europe', 'Asia (Sans Japan)', 'Middle East and Africa', 'Latin America', 'China', 'India', 'EU', 'USA'],
-        [0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000], 									
-        [3104.47059, 0.00000, 8.48186, 2065.81780, 852.00259, 178.16835, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3161.44344, 0.00000, 8.68020, 2095.03147, 876.99695, 180.73482, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3220.03211, 0.00000, 8.88431, 2124.97121, 902.80900, 183.36760, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3280.01976, 0.00000, 9.09242, 2155.47220, 929.40282, 186.05232, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3341.07264, 0.00000, 9.30224, 2186.28270, 956.72112, 188.76657, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3402.93540, 0.00000, 9.51196, 2217.20758, 984.72307, 191.49280, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3465.51278, 0.00000, 9.72091, 2248.17352, 1013.39323, 194.22513, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3528.83717, 0.00000, 9.92908, 2279.19598, 1042.74746, 196.96466, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3592.92607, 0.00000, 10.13609, 2310.26903, 1072.81245, 199.70851, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3657.83624, 0.00000, 10.34174, 2341.41128, 1103.62806, 202.45516, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3723.58350, 0.00000, 10.54590, 2372.61663, 1135.21914, 205.20184, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3790.16283, 0.00000, 10.74844, 2403.87092, 1167.59705, 207.94643, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3857.46641, 0.00000, 10.94930, 2435.09384, 1200.74062, 210.68266, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3925.28660, 0.00000, 11.14859, 2466.13620, 1234.60293, 213.39888, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3993.35721, 0.00000, 11.34656, 2496.81047, 1269.11927, 216.08091, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4061.48558, 0.00000, 11.54352, 2526.98012, 1304.24345, 218.71849, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4129.59718, 0.00000, 11.73953, 2556.59160, 1339.95944, 221.30662, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4197.73264, 0.00000, 11.93500, 2585.66899, 1376.28164, 223.84701, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4265.96225, 0.00000, 12.13112, 2614.25436, 1413.23206, 226.34472, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4334.41009, 0.00000, 12.32938, 2642.42615, 1450.84617, 228.80839, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4403.18004, 0.00000, 12.53094, 2670.25277, 1489.15146, 231.24487, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4472.28136, 0.00000, 12.73618, 2697.73783, 1528.15208, 233.65528, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4541.71837, 0.00000, 12.94510, 2724.89050, 1567.84374, 236.03903, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4611.58966, 0.00000, 13.15806, 2751.79605, 1608.23606, 238.39949, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4682.01590, 0.00000, 13.37532, 2778.56113, 1649.33899, 240.74046, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4753.08948, 0.00000, 13.59698, 2805.26766, 1691.15979, 243.06505, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4824.84801, 0.00000, 13.82315, 2831.94860, 1733.70139, 245.37486, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4897.29233, 0.00000, 14.05360, 2858.60440, 1776.96353, 247.67080, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4970.42889, 0.00000, 14.28755, 2885.24173, 1820.94535, 249.95426, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5044.24933, 0.00000, 14.52396, 2911.85491, 1865.64411, 252.22635, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5118.74120, 0.00000, 14.76191, 2938.43567, 1911.05584, 254.48779, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5193.90327, 0.00000, 15.00109, 2964.98689, 1957.17632, 256.73897, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5269.72440, 0.00000, 15.24128, 2991.50390, 2003.99958, 258.97965, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5346.16451, 0.00000, 15.48174, 3017.95737, 2051.51664, 261.20876, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5423.17037, 0.00000, 15.72168, 3044.30739, 2099.71659, 263.42471, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5500.69146, 0.00000, 15.96048, 3070.51724, 2148.58788, 265.62586, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5578.70692, 0.00000, 16.19784, 3096.57572, 2198.12176, 267.81160, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5657.18672, 0.00000, 16.43364, 3122.46534, 2248.30692, 269.98082, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5736.05888, 0.00000, 16.66779, 3148.13496, 2299.12556, 272.13057, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5815.23808, 0.00000, 16.90025, 3173.52337, 2350.55718, 274.25728, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5894.65704, 0.00000, 17.13109, 3198.58449, 2402.58342, 276.35803, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5974.28702, 0.00000, 17.36026, 3223.30280, 2455.19229, 278.43168, 0.00000, 0.00000, 0.00000, 0.00000],
-        [6054.11935, 0.00000, 17.58787, 3247.68046, 2508.37334, 280.47767, 0.00000, 0.00000, 0.00000, 0.00000],
-        [6134.13467, 0.00000, 17.81440, 3271.71389, 2562.11186, 282.49452, 0.00000, 0.00000, 0.00000, 0.00000],
-        [6214.32057, 0.00000, 18.04046, 3295.40651, 2616.39280, 284.48080, 0.00000, 0.00000, 0.00000, 0.00000],
-        [6294.67271, 0.00000, 18.26660, 3318.76742, 2671.20309, 286.43561, 0.00000, 0.00000, 0.00000, 0.00000]]
-
-# Population in Countries with <0.96 gender partiy in completing upper secondary education, by IPCC Region, REF2_population_scenario (millions)
-# Population_Tables!O61:X108
-ref2_low_edu_list = [
-        ['World', 'OECD90', 'Eastern Europe', 'Asia (Sans Japan)', 'Middle East and Africa', 'Latin America', 'China', 'India', 'EU', 'USA'],
-        [0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000], 
-        [3104.47059, 0.00000, 8.48186, 2065.81780, 852.00259, 178.16835, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3157.87084, 0.00000, 8.66946, 2092.57300, 876.10862, 180.51975, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3211.30766, 0.00000, 8.85812, 2118.97437, 900.63206, 182.84313, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3264.76518, 0.00000, 9.04672, 2145.00457, 925.57677, 185.13711, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3318.25114, 0.00000, 9.23406, 2170.65849, 950.95757, 187.40103, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3371.76847, 0.00000, 9.41913, 2195.92908, 976.78637, 189.63390, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3425.28557, 0.00000, 9.60148, 2220.79237, 1003.05767, 191.83405, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3478.77262, 0.00000, 9.78092, 2245.22595, 1029.76640, 193.99934, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3532.24074, 0.00000, 9.95717, 2269.22863, 1056.92708, 196.12787, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3585.71316, 0.00000, 10.13005, 2292.80580, 1084.55961, 198.21770, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3639.20392, 0.00000, 10.29952, 2315.95853, 1112.67872, 200.26714, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3692.70437, 0.00000, 10.46547, 2338.67720, 1141.28675, 202.27495, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3746.19506, 0.00000, 10.62803, 2360.94646, 1170.38044, 204.24013, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3799.66619, 0.00000, 10.78779, 2382.75504, 1199.96176, 206.16160, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3853.10608, 0.00000, 10.94556, 2404.09065, 1230.03146, 208.03841, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3906.50103, 0.00000, 11.10204, 2424.94118, 1260.58802, 209.86979, 0.00000, 0.00000, 0.00000, 0.00000],
-        [3959.84065, 0.00000, 11.25742, 2445.30062, 1291.62773, 211.65488, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4013.10831, 0.00000, 11.41193, 2465.15923, 1323.14393, 213.39321, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4066.27568, 0.00000, 11.56632, 2484.49702, 1355.12731, 215.08503, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4119.30900, 0.00000, 11.72150, 2503.29046, 1387.56614, 216.73090, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4172.17921, 0.00000, 11.87814, 2521.52214, 1420.44772, 218.33121, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4224.86340, 0.00000, 12.03645, 2539.17960, 1453.76163, 219.88573, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4277.34792, 0.00000, 12.19636, 2556.26285, 1487.49472, 221.39399, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4329.62380, 0.00000, 12.35787, 2572.78277, 1521.62736, 222.85580, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4381.68639, 0.00000, 12.52081, 2588.75763, 1556.13703, 224.27092, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4433.52631, 0.00000, 12.68498, 2604.19966, 1591.00258, 225.63910, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4485.13066, 0.00000, 12.85032, 2619.11179, 1626.20838, 226.96017, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4536.47563, 0.00000, 13.01656, 2633.48566, 1661.73955, 228.23386, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4587.52794, 0.00000, 13.18285, 2647.30893, 1697.57640, 229.45976, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4638.24828, 0.00000, 13.34810, 2660.56442, 1733.69837, 230.63739, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4688.60419, 0.00000, 13.51141, 2673.24003, 1770.08642, 231.76633, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4738.57322, 0.00000, 13.67245, 2685.33108, 1806.72343, 232.84627, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4788.14492, 0.00000, 13.83102, 2696.84193, 1843.59496, 233.87701, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4837.31373, 0.00000, 13.98669, 2707.78058, 1880.68803, 234.85843, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4886.07986, 0.00000, 14.13900, 2718.15927, 1917.99112, 235.79048, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4934.43967, 0.00000, 14.28761, 2727.98702, 1955.49202, 236.67302, 0.00000, 0.00000, 0.00000, 0.00000],
-        [4982.38243, 0.00000, 14.43231, 2737.26795, 1993.17616, 237.50601, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5029.89046, 0.00000, 14.57306, 2745.99994, 2031.02828, 238.28918, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5076.94541, 0.00000, 14.70994, 2754.17866, 2069.03481, 239.02200, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5123.52646, 0.00000, 14.84321, 2761.79715, 2107.18230, 239.70380, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5169.61579, 0.00000, 14.97308, 2768.85153, 2145.45707, 240.33411, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5215.20349, 0.00000, 15.09957, 2775.34612, 2183.84488, 240.91292, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5260.28197, 0.00000, 15.22279, 2781.28761, 2222.33114, 241.44043, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5304.83924, 0.00000, 15.34313, 2786.67810, 2260.90125, 241.91676, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5348.86356, 0.00000, 15.46109, 2791.51990, 2299.54044, 242.34214, 0.00000, 0.00000, 0.00000, 0.00000],
-        [5392.34612, 0.00000, 15.57707, 2795.81825, 2338.23390, 242.71689, 0.00000, 0.00000, 0.00000, 0.00000]]
-
-# Table 2: REF2, Electricity Generation TAM (TWh)										
-# Electricity_cluster!B26:K73
-ref2_tam_list = [
-        ['World', 'OECD90', 'Eastern Europe', 'Asia (Sans Japan)', 'Middle East and Africa', 'Latin America', 'China', 'India', 'EU', 'USA'],
-        [22548.00000, 9630.94132, 2021.81456, 8068.09850, 1750.27240, 1681.57391, 5262.77320, 1324.87968, 3379.55071, 4226.08258],
-        [23915.29474, 9686.06712, 2046.08149, 8518.24480, 1813.43945, 1729.37357, 5577.46645, 1407.75811, 3402.87973, 4238.15940],
-        [24739.34640, 9726.59905, 2070.42874, 8971.48190, 1887.43392, 1782.38592, 5868.27237, 1508.56127, 3423.02929, 4238.12779],
-        [25546.28067, 9770.88424, 2095.77922, 9418.25671, 1964.77842, 1837.98642, 6148.15200, 1613.15158, 3443.15125, 4240.90633],
-        [26336.89072, 9818.79572, 2122.08509, 9858.95275, 2045.58829, 1896.13291, 6417.38657, 1721.48637, 3463.30080, 4246.39976],
-        [27111.96969, 9870.20653, 2149.29851, 10293.95356, 2129.97886, 1956.78323, 6676.25733, 1833.52292, 3483.53313, 4254.51283],
-        [27872.31074, 9924.98972, 2177.37163, 10723.64264, 2218.06547, 2019.89524, 6925.04552, 1949.21857, 3503.90343, 4265.15027],
-        [28618.70702, 9983.01832, 2206.25661, 11148.40352, 2309.96345, 2085.42679, 7164.03237, 2068.53061, 3524.46690, 4278.21684],
-        [29351.95169, 10044.16537, 2235.90560, 11568.61971, 2405.78813, 2153.33573, 7393.49912, 2191.41636, 3545.27873, 4293.61727],
-        [30072.83791, 10108.30392, 2266.27077, 11984.67474, 2505.65485, 2223.57990, 7613.72702, 2317.83312, 3566.39411, 4311.25631],
-        [30782.15882, 10175.30699, 2297.30426, 12396.95213, 2609.67894, 2296.11715, 7824.99730, 2447.73821, 3587.86825, 4331.03871],
-        [31480.70758, 10245.04764, 2328.95824, 12805.83539, 2717.97573, 2370.90533, 8027.59121, 2581.08893, 3609.75633, 4352.86921],
-        [32169.27735, 10317.39891, 2361.18485, 13211.70805, 2830.66056, 2447.90230, 8221.78998, 2717.84260, 3632.11354, 4376.65255],
-        [32848.66128, 10392.23382, 2393.93626, 13614.95362, 2947.84876, 2527.06589, 8407.87486, 2857.95652, 3654.99508, 4402.29347],
-        [33519.65253, 10469.42543, 2427.16462, 14015.95562, 3069.65567, 2608.35396, 8586.12709, 3001.38800, 3678.45614, 4429.69673],
-        [34183.04424, 10548.84677, 2460.82209, 14415.09758, 3196.19661, 2691.72436, 8756.82789, 3148.09436, 3702.55192, 4458.76706],
-        [34839.62958, 10630.37088, 2494.86082, 14812.76302, 3327.58693, 2777.13494, 8920.25852, 3298.03290, 3727.33762, 4489.40921],
-        [35490.20169, 10713.87080, 2529.23298, 15209.33545, 3463.94195, 2864.54354, 9076.70022, 3451.16094, 3752.86841, 4521.52793],
-        [36135.55373, 10799.21958, 2563.89071, 15605.19839, 3605.37701, 2953.90802, 9226.43422, 3607.43578, 3779.19950, 4555.02795],
-        [36776.47886, 10886.29024, 2598.78617, 16000.73536, 3752.00744, 3045.18622, 9369.74177, 3766.81473, 3806.38608, 4589.81402],
-        [37413.77023, 10974.95584, 2633.87152, 16396.32989, 3903.94858, 3138.33599, 9506.90409, 3929.25510, 3834.48334, 4625.79089],
-        [38048.22099, 11065.08941, 2669.09891, 16792.36549, 4061.31575, 3233.31518, 9638.20245, 4094.71421, 3863.54649, 4662.86330],
-        [38680.62430, 11156.56399, 2704.42051, 17189.22569, 4224.22430, 3330.08164, 9763.91807, 4263.14936, 3893.63070, 4700.93599],
-        [39311.77332, 11249.25262, 2739.78846, 17587.29399, 4392.78956, 3428.59321, 9884.33219, 4434.51785, 3924.79118, 4739.91371],
-        [39942.46119, 11343.02835, 2775.15493, 17986.95393, 4567.12685, 3528.80776, 9999.72605, 4608.77701, 3957.08311, 4779.70121],
-        [40573.48107, 11437.76420, 2810.47206, 18388.58902, 4747.35152, 3630.68312, 10110.38090, 4785.88414, 3990.56170, 4820.20321],
-        [41205.62612, 11533.33323, 2845.69203, 18792.58278, 4933.57890, 3734.17715, 10216.57798, 4965.79655, 4025.28213, 4861.32448],
-        [41839.68949, 11629.60846, 2880.76697, 19199.31873, 5125.92432, 3839.24769, 10318.59852, 5148.47155, 4061.29960, 4902.96975],
-        [42476.46433, 11726.46295, 2915.64905, 19609.18039, 5324.50311, 3945.85259, 10416.72376, 5333.86645, 4098.66931, 4945.04377],
-        [43116.74380, 11823.76973, 2950.29042, 20022.55128, 5529.43060, 4053.94970, 10511.23495, 5521.93856, 4137.44643, 4987.45129],
-        [43761.32106, 11921.40184, 2984.64325, 20439.81493, 5740.82214, 4163.49688, 10602.41332, 5712.64519, 4177.68618, 5030.09703],
-        [44410.98925, 12019.23231, 3018.65968, 20861.35484, 5958.79306, 4274.45196, 10690.54011, 5905.94365, 4219.44374, 5072.88576],
-        [45066.54153, 12117.13420, 3052.29187, 21287.55454, 6183.45868, 4386.77280, 10775.89657, 6101.79125, 4262.77431, 5115.72222],
-        [45728.77106, 12214.98054, 3085.49198, 21718.79755, 6414.93434, 4500.41725, 10858.76393, 6300.14529, 4307.73308, 5158.51114],
-        [46398.47098, 12312.64437, 3118.21217, 22155.46740, 6653.33538, 4615.34315, 10939.42343, 6500.96309, 4354.37524, 5201.15727],
-        [47076.43447, 12409.99873, 3150.40458, 22597.94759, 6898.77713, 4731.50836, 11018.15632, 6704.20196, 4402.75599, 5243.56536],
-        [47763.45466, 12506.91666, 3182.02139, 23046.62165, 7151.37491, 4848.87072, 11095.24383, 6909.81921, 4452.93051, 5285.64015],
-        [48460.32471, 12603.27120, 3213.01473, 23501.87309, 7411.24408, 4967.38809, 11170.96721, 7117.77215, 4504.95402, 5327.28639],
-        [49167.83778, 12698.93539, 3243.33678, 23964.08545, 7678.49995, 5087.01830, 11245.60769, 7328.01808, 4558.88169, 5368.40881],
-        [49886.78702, 12793.78227, 3272.93968, 24433.64223, 7953.25787, 5207.71922, 11319.44651, 7540.51431, 4614.76872, 5408.91217],
-        [50617.96559, 12887.68488, 3301.77559, 24910.92696, 8235.63316, 5329.44868, 11392.76492, 7755.21817, 4672.67030, 5448.70121],
-        [51362.16663, 12980.51625, 3329.79667, 25396.32316, 8525.74117, 5452.16454, 11465.84415, 7972.08695, 4732.64164, 5487.68066],
-        [52120.18332, 13072.14944, 3356.95508, 25890.21435, 8823.69721, 5575.82465, 11538.96544, 8191.07796, 4794.73791, 5525.75528],
-        [52892.80878, 13162.45748, 3383.20296, 26392.98404, 9129.61663, 5700.38686, 11612.41004, 8412.14852, 4859.01432, 5562.82981],
-        [53680.83620, 13251.31340, 3408.49248, 26905.01576, 9443.61477, 5825.80901, 11686.45918, 8635.25594, 4925.52606, 5598.80900],
-        [54485.05871, 13338.59025, 3432.77580, 27426.69302, 9765.80695, 5952.04895, 11761.39410, 8860.35752, 4994.32832, 5633.59758],
-        [55306.26947, 13424.16108, 3456.00506, 27958.39935, 10096.30850, 6079.06453, 11837.49604, 9087.41057, 5065.47630, 5667.10030]]
-
