--- solution/healthandeducation/test_electricity.py
@@ -1,10 +1,0 @@
-import sys
-sys.path.append('c:\\Users\\sunishchal.dev\\Documents\\solutions\\solution')
-
-import healthandeducation as he
-
-h = he.Scenario()
-
-print(h.ref2_tam.tail())
-print(h.ref1_tam_all_regions.tail())
-print(h.change_elec_gen.tail())
